Add files via upload
This project aims to predict the eligibility of customers for loan approval using the "Loan_Data.csv" dataset, which has 614 values and 13 columns. To do this, we will analyse the data, process it, balance the classes, and then apply machine learning classification models. Through this process, we can determine if customers are eligible or not for loan approval.